requirements.txt
pycryptodome~=3.11.0
Flask~=2.0.2
waitress~=2.0.0